+++ lib/core/database/adapters/lunasea.dart
@@ -11,10 +11,11 @@ class LunaSeaDatabase extends LunaModuleDatabase {
         Map<String, dynamic> data = {};
         for(LunaSeaDatabaseValue value in LunaSeaDatabaseValue.values) {
             switch(value) {
-                case LunaSeaDatabaseValue.SELECTED_BROWSER: data[value.key] = (LunaSeaDatabaseValue.SELECTED_BROWSER as LSBrowsers).key; break;
+                // Non-primitive values
+                case LunaSeaDatabaseValue.SELECTED_BROWSER: data[value.key] = (LunaSeaDatabaseValue.SELECTED_BROWSER.data as LSBrowsers).key; break;
+                // Primitive values
                 case LunaSeaDatabaseValue.ENABLED_PROFILE: 
                 case LunaSeaDatabaseValue.ENABLED_SENTRY:
-                case LunaSeaDatabaseValue.CLIENT_IDENTIFIER:
                 case LunaSeaDatabaseValue.THEME_AMOLED:
                 case LunaSeaDatabaseValue.THEME_AMOLED_BORDER:
                 case LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY:
@@ -27,7 +28,6 @@ class LunaSeaDatabase extends LunaModuleDatabase {
                 case LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR:
                 case LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET:
                 case LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD:
-                case LunaSeaDatabaseValue.QUICK_ACTIONS_OMBI:
                 case LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI:
                 case LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH:
                 case LunaSeaDatabaseValue.USE_24_HOUR_TIME: data[value.key] = value.data;
@@ -38,14 +38,61 @@ class LunaSeaDatabase extends LunaModuleDatabase {
 
     @override
     void import(Map<String, dynamic> config) {
-        // TODO
+        for(String key in config.keys) {
+            LunaSeaDatabaseValue value = valueFromKey(key);
+            if(value != null) switch(value) {
+                // Non-primitive values
+                case LunaSeaDatabaseValue.SELECTED_BROWSER: value.put(LSBrowsers.APPLE_SAFARI.fromKey(config[key])); break;
+                // Primitive values
+                case LunaSeaDatabaseValue.ENABLED_PROFILE: 
+                case LunaSeaDatabaseValue.ENABLED_SENTRY:
+                case LunaSeaDatabaseValue.THEME_AMOLED:
+                case LunaSeaDatabaseValue.THEME_AMOLED_BORDER:
+                case LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY:
+                case LunaSeaDatabaseValue.DRAWER_GROUP_MODULES:
+                case LunaSeaDatabaseValue.DRAWER_EXPAND_AUTOMATION:
+                case LunaSeaDatabaseValue.DRAWER_EXPAND_CLIENTS:
+                case LunaSeaDatabaseValue.DRAWER_EXPAND_MONITORING:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_LIDARR:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_RADARR:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI:
+                case LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH:
+                case LunaSeaDatabaseValue.USE_24_HOUR_TIME: value.put(config[key]); break;
+            }
+        }
+    }
+
+    LunaSeaDatabaseValue valueFromKey(String key) {
+        switch(key) {
+            case 'profile': return LunaSeaDatabaseValue.ENABLED_PROFILE;
+            case 'LUNASEA_ENABLED_SENTRY': return LunaSeaDatabaseValue.ENABLED_SENTRY;
+            case 'LUNASEA_THEME_AMOLED': return LunaSeaDatabaseValue.THEME_AMOLED;
+            case 'LUNASEA_THEME_AMOLED_BORDER': return LunaSeaDatabaseValue.THEME_AMOLED_BORDER;
+            case 'LUNASEA_THEME_IMAGE_BACKGROUND_OPACITY': return LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY;
+            case 'LUNASEA_SELECTED_BROWSER': return LunaSeaDatabaseValue.SELECTED_BROWSER;
+            case 'LUNASEA_DRAWER_GROUP_MODULES': return LunaSeaDatabaseValue.DRAWER_GROUP_MODULES;
+            case 'LUNASEA_DRAWER_EXPAND_AUTOMATION': return LunaSeaDatabaseValue.DRAWER_EXPAND_AUTOMATION;
+            case 'LUNASEA_DRAWER_EXPAND_CLIENTS': return LunaSeaDatabaseValue.DRAWER_EXPAND_CLIENTS;
+            case 'LUNASEA_DRAWER_EXPAND_MONITORING': return LunaSeaDatabaseValue.DRAWER_EXPAND_MONITORING;
+            case 'LUNASEA_QUICK_ACTIONS_LIDARR': return LunaSeaDatabaseValue.QUICK_ACTIONS_LIDARR;
+            case 'LUNASEA_QUICK_ACTIONS_RADARR': return LunaSeaDatabaseValue.QUICK_ACTIONS_RADARR;
+            case 'LUNASEA_QUICK_ACTIONS_SONARR': return LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR;
+            case 'LUNASEA_QUICK_ACTIONS_NZBGET': return LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET;
+            case 'LUNASEA_QUICK_ACTIONS_SABNZBD': return LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD;
+            case 'LUNASEA_QUICK_ACTIONS_TAUTULLI': return LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI;
+            case 'LUNASEA_QUICK_ACTIONS_SEARCH': return LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH;
+            case 'LUNASEA_USE_24_HOUR_TIME': return LunaSeaDatabaseValue.USE_24_HOUR_TIME;
+        }
+        return null;
     }
 }
 
 enum LunaSeaDatabaseValue {
     ENABLED_PROFILE,
     ENABLED_SENTRY,
-    CLIENT_IDENTIFIER,
     THEME_AMOLED,
     THEME_AMOLED_BORDER,
     THEME_IMAGE_BACKGROUND_OPACITY,
@@ -59,7 +106,6 @@ enum LunaSeaDatabaseValue {
     QUICK_ACTIONS_SONARR,
     QUICK_ACTIONS_NZBGET,
     QUICK_ACTIONS_SABNZBD,
-    QUICK_ACTIONS_OMBI,
     QUICK_ACTIONS_TAUTULLI,
     QUICK_ACTIONS_SEARCH,
     USE_24_HOUR_TIME,
@@ -74,7 +120,6 @@ extension LunaSeaDatabaseValueExtension on LunaSeaDatabaseValue {
             case LunaSeaDatabaseValue.THEME_AMOLED_BORDER: return 'LUNASEA_THEME_AMOLED_BORDER';
             case LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY: return 'LUNASEA_THEME_IMAGE_BACKGROUND_OPACITY';
             case LunaSeaDatabaseValue.SELECTED_BROWSER: return 'LUNASEA_SELECTED_BROWSER';
-            case LunaSeaDatabaseValue.CLIENT_IDENTIFIER: return 'LUNASEA_CLIENT_IDENTIFIER';
             case LunaSeaDatabaseValue.DRAWER_GROUP_MODULES: return 'LUNASEA_DRAWER_GROUP_MODULES';
             case LunaSeaDatabaseValue.DRAWER_EXPAND_AUTOMATION: return 'LUNASEA_DRAWER_EXPAND_AUTOMATION';
             case LunaSeaDatabaseValue.DRAWER_EXPAND_CLIENTS: return 'LUNASEA_DRAWER_EXPAND_CLIENTS';
@@ -84,38 +129,11 @@ extension LunaSeaDatabaseValueExtension on LunaSeaDatabaseValue {
             case LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR: return 'LUNASEA_QUICK_ACTIONS_SONARR';
             case LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET: return 'LUNASEA_QUICK_ACTIONS_NZBGET';
             case LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD: return 'LUNASEA_QUICK_ACTIONS_SABNZBD';
-            case LunaSeaDatabaseValue.QUICK_ACTIONS_OMBI: return 'LUNASEA_QUICK_ACTIONS_OMBI';
             case LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI: return 'LUNASEA_QUICK_ACTIONS_TAUTULLI';
             case LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH: return 'LUNASEA_QUICK_ACTIONS_SEARCH';
             case LunaSeaDatabaseValue.USE_24_HOUR_TIME: return 'LUNASEA_USE_24_HOUR_TIME';
         }
         throw Exception('key not found'); 
-    }
-
-    LunaSeaDatabaseValue fromKey(String key) {
-        switch(key) {
-            case 'profile': return LunaSeaDatabaseValue.ENABLED_PROFILE;
-            case 'LUNASEA_ENABLED_SENTRY': return LunaSeaDatabaseValue.ENABLED_SENTRY;
-            case 'LUNASEA_THEME_AMOLED': return LunaSeaDatabaseValue.THEME_AMOLED;
-            case 'LUNASEA_THEME_AMOLED_BORDER': return LunaSeaDatabaseValue.THEME_AMOLED_BORDER;
-            case 'LUNASEA_THEME_IMAGE_BACKGROUND_OPACITY': return LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY;
-            case 'LUNASEA_SELECTED_BROWSER': return LunaSeaDatabaseValue.SELECTED_BROWSER;
-            case 'LUNASEA_CLIENT_IDENTIFIER': return LunaSeaDatabaseValue.CLIENT_IDENTIFIER;
-            case 'LUNASEA_DRAWER_GROUP_MODULES': return LunaSeaDatabaseValue.DRAWER_GROUP_MODULES;
-            case 'LUNASEA_DRAWER_EXPAND_AUTOMATION': return LunaSeaDatabaseValue.DRAWER_EXPAND_AUTOMATION;
-            case 'LUNASEA_DRAWER_EXPAND_CLIENTS': return LunaSeaDatabaseValue.DRAWER_EXPAND_CLIENTS;
-            case 'LUNASEA_DRAWER_EXPAND_MONITORING': return LunaSeaDatabaseValue.DRAWER_EXPAND_MONITORING;
-            case 'LUNASEA_QUICK_ACTIONS_LIDARR': return LunaSeaDatabaseValue.QUICK_ACTIONS_LIDARR;
-            case 'LUNASEA_QUICK_ACTIONS_RADARR': return LunaSeaDatabaseValue.QUICK_ACTIONS_RADARR;
-            case 'LUNASEA_QUICK_ACTIONS_SONARR': return LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR;
-            case 'LUNASEA_QUICK_ACTIONS_NZBGET': return LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET;
-            case 'LUNASEA_QUICK_ACTIONS_SABNZBD': return LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD;
-            case 'LUNASEA_QUICK_ACTIONS_OMBI': return LunaSeaDatabaseValue.QUICK_ACTIONS_OMBI;
-            case 'LUNASEA_QUICK_ACTIONS_TAUTULLI': return LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI;
-            case 'LUNASEA_QUICK_ACTIONS_SEARCH': return LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH;
-            case 'LUNASEA_USE_24_HOUR_TIME': return LunaSeaDatabaseValue.USE_24_HOUR_TIME;
-        }
-        return null;
     }
 
     dynamic get data {
@@ -127,7 +145,6 @@ extension LunaSeaDatabaseValueExtension on LunaSeaDatabaseValue {
             case LunaSeaDatabaseValue.THEME_AMOLED_BORDER: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY: return _box.get(this.key, defaultValue: 10);
             case LunaSeaDatabaseValue.ENABLED_SENTRY: return _box.get(this.key, defaultValue: true);
-            case LunaSeaDatabaseValue.CLIENT_IDENTIFIER: return _box.get(this.key, defaultValue: null);
             case LunaSeaDatabaseValue.DRAWER_GROUP_MODULES: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.DRAWER_EXPAND_AUTOMATION: return _box.get(this.key, defaultValue: true);
             case LunaSeaDatabaseValue.DRAWER_EXPAND_CLIENTS: return _box.get(this.key, defaultValue: true);
@@ -137,7 +154,6 @@ extension LunaSeaDatabaseValueExtension on LunaSeaDatabaseValue {
             case LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD: return _box.get(this.key, defaultValue: false);
-            case LunaSeaDatabaseValue.QUICK_ACTIONS_OMBI: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH: return _box.get(this.key, defaultValue: false);
             case LunaSeaDatabaseValue.USE_24_HOUR_TIME: return _box.get(this.key, defaultValue: false);
@@ -145,5 +161,28 @@ extension LunaSeaDatabaseValueExtension on LunaSeaDatabaseValue {
         throw Exception('data not found'); 
     }
 
-    void put(dynamic value) => Database.lunaSeaBox.put(this.key, value);
+    void put(dynamic value) {
+        final _box = Database.lunaSeaBox;
+        switch(this) {
+            case LunaSeaDatabaseValue.ENABLED_PROFILE: if(value.runtimeType == String) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.ENABLED_SENTRY: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.THEME_AMOLED: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.THEME_AMOLED_BORDER: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.THEME_IMAGE_BACKGROUND_OPACITY: if(value.runtimeType == int) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.SELECTED_BROWSER: if(value.runtimeType == LSBrowsers) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.DRAWER_GROUP_MODULES: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.DRAWER_EXPAND_AUTOMATION: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.DRAWER_EXPAND_CLIENTS: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.DRAWER_EXPAND_MONITORING: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_LIDARR: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_RADARR: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_SONARR: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_NZBGET: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_SABNZBD: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_TAUTULLI: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.QUICK_ACTIONS_SEARCH: if(value.runtimeType == bool) _box.put(this.key, value); return;
+            case LunaSeaDatabaseValue.USE_24_HOUR_TIME: if(value.runtimeType == bool) _box.put(this.key, value); return;
+        }
+        LunaLogger.warning('LunaSeaDatabaseValueExtension', 'put', 'Attempted to enter data for invalid LunaSeaDatabaseValue: ${this?.toString() ?? 'null'}');
+    }
 }