+++ lib/core/database/module_database.dart
@@ -7,4 +7,7 @@ abstract class LunaModuleDatabase {
     
     /// Receiving a map object, import a configuration
     void import(Map<String, dynamic> config);
+
+    /// Return the Database enumerator value given the key
+    dynamic valueFromKey(String value);
 }