+++ lib/core/theme.dart
@@ -22,13 +22,7 @@ class LunaTheme {
             cardColor: LunaColours.secondary,
             splashColor: LunaColours.splash,
             dialogBackgroundColor: LunaColours.secondary,
-            dividerColor: LunaColours.accent.withAlpha(0),
             toggleableActiveColor: LunaColours.accent,
-            dividerTheme: DividerThemeData(
-                color: LunaColours.accent,
-                indent: 100.0,
-                endIndent: 100.0,
-            ),
             iconTheme: IconThemeData(
                 color: Colors.white,
             ),
@@ -65,13 +59,7 @@ class LunaTheme {
             cardColor: Colors.black,
             splashColor: LunaColours.splash,
             dialogBackgroundColor: Colors.black,
-            dividerColor: LunaColours.accent.withAlpha(0),
             toggleableActiveColor: LunaColours.accent,
-            dividerTheme: DividerThemeData(
-                color: LunaColours.accent,
-                indent: 72.0,
-                endIndent: 72.0,
-            ),
             iconTheme: IconThemeData(
                 color: Colors.white,
             ),